refactor:optimize auto configuration.
