+++ spring-cloud-starter-tencent-polaris-auth/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.tencent.cloud.polaris.auth.config.PolarisAuthPropertiesAutoConfiguration,\
   com.tencent.cloud.polaris.auth.config.PolarisAuthAutoConfiguration
-org.springframework.cloud.bootstrap.BootstrapConfiguration=\
-  com.tencent.cloud.polaris.auth.config.PolarisAuthPropertiesBootstrapConfiguration
