--- spring-cloud-starter-tencent-polaris-discovery/src/test/java/com/tencent/cloud/polaris/DiscoveryPropertiesBootstrapAutoConfigurationTest.java
@@ -1,47 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making spring-cloud-tencent available.
- *
- * Copyright (C) 2021 Tencent. All rights reserved.
- *
- * Licensed under the BSD 3-Clause License (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://opensource.org/licenses/BSD-3-Clause
- *
- * Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package com.tencent.cloud.polaris;
-
-import com.tencent.cloud.polaris.context.config.PolarisContextAutoConfiguration;
-import org.junit.jupiter.api.Test;
-
-import org.springframework.boot.autoconfigure.AutoConfigurations;
-import org.springframework.boot.test.context.runner.ApplicationContextRunner;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-/**
- * Test for {@link DiscoveryPropertiesBootstrapAutoConfiguration}.
- *
- * @author Haotian Zhang
- */
-public class DiscoveryPropertiesBootstrapAutoConfigurationTest {
-
-	@Test
-	public void testDefaultInitialization() {
-		ApplicationContextRunner applicationContextRunner = new ApplicationContextRunner().withConfiguration(
-						AutoConfigurations.of(
-								PolarisContextAutoConfiguration.class,
-								DiscoveryPropertiesBootstrapAutoConfiguration.class))
-				.withPropertyValues("spring.cloud.polaris.enabled=true");
-		applicationContextRunner.run(context -> {
-			assertThat(context).hasSingleBean(DiscoveryPropertiesBootstrapAutoConfiguration.class);
-			assertThat(context).hasSingleBean(DiscoveryPropertiesAutoConfiguration.class);
-		});
-	}
-}
