+++ spring-cloud-starter-tencent-polaris-router/src/main/resources/META-INF/spring.factories
@@ -3,5 +3,3 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.tencent.cloud.polaris.router.config.RouterAutoConfiguration,\
   com.tencent.cloud.polaris.router.config.FeignAutoConfiguration,\
   com.tencent.cloud.polaris.router.endpoint.PolarisRouterEndpointAutoConfiguration
-org.springframework.cloud.bootstrap.BootstrapConfiguration=\
-  com.tencent.cloud.polaris.router.config.RouterBootstrapAutoConfiguration
