+++ spring-cloud-tencent-plugin-starters/spring-cloud-starter-tencent-trace-plugin/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.tencent.cloud.plugin.trace.config.TraceConfigModifierAutoConfiguration,\
   com.tencent.cloud.plugin.trace.config.TraceEnhancedPluginAutoConfiguration
-org.springframework.cloud.bootstrap.BootstrapConfiguration=\
-  com.tencent.cloud.plugin.trace.config.TraceConfigModifierBootstrapAutoConfiguration
