--- spring-cloud-tencent-plugin-starters/spring-cloud-tencent-lossless-plugin/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,10 +1,0 @@
-{
-  "properties": [
-    {
-      "name": "spring.cloud.polaris.lossless.enabled",
-      "type": "java.lang.Boolean",
-      "defaultValue": false,
-      "description": "the switch for lossless plugin."
-    }
-  ]
-}
