+++ spring-cloud-tencent-plugin-starters/spring-cloud-tencent-lossless-plugin/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,2 @@
-org.springframework.cloud.bootstrap.BootstrapConfiguration=\
-  com.tencent.cloud.plugin.lossless.config.LosslessPropertiesBootstrapConfiguration
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.tencent.cloud.plugin.lossless.config.LosslessAutoConfiguration
