+++ spring-cloud-tencent-polaris-context/src/main/resources/META-INF/spring.factories
@@ -3,8 +3,7 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.tencent.cloud.polaris.context.config.PolarisContextPostConfiguration,\
   com.tencent.cloud.polaris.context.config.extend.tsf.TsfContextAutoConfiguration
 org.springframework.cloud.bootstrap.BootstrapConfiguration=\
-  com.tencent.cloud.polaris.context.config.PolarisContextBootstrapAutoConfiguration,\
-  com.tencent.cloud.polaris.context.config.extend.tsf.TsfContextBootstrapConfiguration
+    com.tencent.cloud.polaris.context.config.PolarisContextPropertiesBootstrapAutoConfiguration
 org.springframework.context.ApplicationListener=\
   com.tencent.cloud.polaris.context.logging.PolarisLoggingApplicationListener,\
   com.tencent.cloud.polaris.context.listener.FailedEventApplicationListener
