+++ spring-cloud-tencent-rpc-enhancement/src/main/resources/META-INF/spring.factories
@@ -1,7 +1,3 @@
-org.springframework.cloud.bootstrap.BootstrapConfiguration=\
-  com.tencent.cloud.rpc.enhancement.stat.config.PolarisStatPropertiesBootstrapConfiguration,\
-  com.tencent.cloud.rpc.enhancement.config.RpcEnhancementBootstrapConfiguration,\
-  com.tencent.cloud.rpc.enhancement.config.RpcEnhancementPropertiesBootstrapConfiguration
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.tencent.cloud.rpc.enhancement.config.RpcEnhancementAutoConfiguration,\
   com.tencent.cloud.rpc.enhancement.stat.config.PolarisStatPropertiesAutoConfiguration,\
